Add confirm alert on contact deletion

+++ src/components/ContactList/ContactList.jsx
@@ -1,7 +1,19 @@
 import { nanoid } from 'nanoid';
-import { ContactInfo, ListBtn, ListItem, ListWrapper } from './ContactList.styled';
+import {
+  BtnWrapper,
+  ContactInfo,
+  ListBtn,
+  ListItem,
+  ListWrapper,
+  ModalBtn,
+  ModalTitle,
+  ModalWrapper,
+} from './ContactList.styled';
 import { useSelector, useDispatch } from 'react-redux';
 import { deleteContact } from 'redux/contacts/contactsOperations';
+import { confirmAlert } from 'react-confirm-alert';
+import 'react-confirm-alert/src/react-confirm-alert.css';
+import {  toast } from 'react-toastify';
 
 export const ContactList = () => {
   const dispatch = useDispatch();
@@ -9,9 +21,42 @@ export const ContactList = () => {
   const contacts = useSelector(state => state.contacts.contacts);
   const filter = useSelector(state => state.filter.filter);
 
+  const submit = id => {
+    confirmAlert({
+      customUI: ({ onClose }) => {
+        return (
+          <ModalWrapper className="custom-ui">
+            <ModalTitle>Are you sure?</ModalTitle>
+            <p>You want to delete this contact?</p>
+            <BtnWrapper>
+              <ModalBtn onClick={onClose}>No</ModalBtn>
+              <ModalBtn
+                onClick={() => {
+                  dispatch(deleteContact(id));
+                  toast.success('Success! The contact has been deleted', {
+                    position: 'top-right',
+                    autoClose: 3000,
+                    hideProgressBar: false,
+                    closeOnClick: true,
+                    pauseOnHover: true,
+                    draggable: true,
+                    progress: undefined,
+                    theme: 'light',
+                  });
+                  onClose();
+                }}
+              >
+                Yes!
+              </ModalBtn>
+            </BtnWrapper>
+          </ModalWrapper>
+        );
+      },
+    });
+  };
+
   const handleDeleteContact = id => {
-    
-    dispatch(deleteContact(id));
+    submit(id);
   };
 
   const normalizeFilter = filter.toLocaleLowerCase();
